requirements.txt
opencv-python
tensorflow
ultralytics
pillow
os
numpy
matplotlib.pyplot